+++ src/lib/update/plane.c
@@ -182,8 +182,8 @@ prod_plane(int etus, int natnum, int *bp, int buildem)
 	    delta = roundavg((double)avail / w_p_eff);
 	    if (delta <= 0)
 		continue;
-	    if (delta > etus * plane_grow_scale)
-		delta = etus * plane_grow_scale;
+	    if (delta > (int)((float)etus * plane_grow_scale))
+		delta = (int)((float)etus * plane_grow_scale);
 	    if (delta > left)
 		delta = left;
